Fixed the latitude and longitude and changed the price to integer

+++ database/migrations/2017_06_23_182455_create_properties_table.php
@@ -20,12 +20,12 @@ class CreatePropertiesTable extends Migration
 		$table->string('number')->nullable();
 		$table->string('floor')->nullable();
 		$table->string('type');
-		$table->float('price');
+		$table->integer('price');
 		$table->integer('capacity')->nullable();
 		$table->string('availibility')->nullable();
 		$table->integer('number_wcs')->nullable();
-		$table->float('latitude')->nullable();
-		$table->float('longitude')->nullable();
+		$table->float('latitude',6)->nullable();
+		$table->float('longitude',6)->nullable();
 		$table->boolean('has_living_room')->nullable();
 		$table->boolean('has_cleaning')->nullable();
 		$table->boolean('expenses_included')->nullable();
